requirements.txt
# =============================================================================
# PetCare Medellín - AI Automation Dependencies
# =============================================================================

# Core
openai>=1.30.0
python-dotenv>=1.0.1

# -----------------------------------------------------------------------------
# RAG - Document Loading & Chunking
# -----------------------------------------------------------------------------
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-ollama>=0.2.0
langchain-text-splitters>=0.3.0
langchain-chroma>=1.0.0

# -----------------------------------------------------------------------------
# Vector Database
# -----------------------------------------------------------------------------
chromadb>=0.6.0

# -----------------------------------------------------------------------------
# Document Parsing
# -----------------------------------------------------------------------------
markdown>=3.6
pypdf>=5.0.0
tiktoken>=0.7.0
unstructured>=0.18.0

# -----------------------------------------------------------------------------
# Testing
# -----------------------------------------------------------------------------
pytest>=8.0.0
pytest-mock>=3.14.0
